move social icons to shared file, add google icon
