+++ _inc/client/scss/variables/_colors.scss
@@ -86,6 +86,7 @@ $color-print:       #f8f8f8;
 $color-skype:       #00AFF0;
 $color-telegram:    #0088cc;
 $color-whatsapp:    #43d854;
+$color-google:      #4285F4;
 
 
 // ==========================================================================
