+++ extensions/blocks/publicize/editor.scss
@@ -24,21 +24,7 @@
 }
 
 .jetpack-publicize-gutenberg-social-icon {
-	fill: $dark-gray-500;
 	margin-right: 5px;
-
-	&.is-facebook {
-		fill: var( --color-facebook );
-	}
-	&.is-twitter {
-		fill: var( --color-twitter );
-	}
-	&.is-linkedin {
-		fill: var( --color-linkedin );
-	}
-	&.is-tumblr {
-		fill: var( --color-tumblr );
-	}
 }
 
 .jetpack-publicize-connection-label {
